shore: Re-structure the project better by moving certificate related parts into a separate folder
